add button work globaly
