+++ app/components/cart-check-list.hbs
@@ -5,11 +5,11 @@
     <div class="name-totals p-6 w-full flex flex-col items-end border border-gray-50">
         <div class="subtotal w-8/12 flex justify-between items-center mb-3">
             <span class="name text-gray-800">Subtotal</span>
-            <span class="price text-gray-800">£{{345}}</span>
+            <span class="price text-gray-800">£{{this.cart.totalCartPrice}}</span>
         </div>
         <div class="subtotal w-8/12 flex justify-between items-center mb-3">
             <span class="name text-gray-800">Shipping Cost</span>
-            <span class="price text-gray-800">£{{345}}</span>
+            <span class="price text-gray-800">£{{0}}</span>
         </div>
         <div class="subtotal w-8/12 flex justify-between items-center mb-3">
             <span class="name text-gray-800">Discount</span>
@@ -19,7 +19,7 @@
     <div class="ammount w-full p-6">
         <div class="total-summ w-full flex justify-between items-center mb-4">
             <span class="name text-xl text-gray-800">Total Payable</span>
-            <span class="price text-xl text-gray-800 font-bold">£{{345}}</span>
+            <span class="price text-xl text-gray-800 font-bold">£{{this.cart.totalCartPrice}}</span>
         </div>
         <button class="bg-violet-500 rounded py-1 px-2 text-white font-bold ml-auto block hover:bg-violet-50 hover:text-violet-500 transition" type="button">Go To Checkout</button>
     </div>
