+++ app/components/cart-check-list.js
@@ -1,7 +1,6 @@
 import Component from '@glimmer/component';
 import { service } from '@ember/service';
-// import { action } from '@ember/object';
 
-export default class CheckoutMainComponent extends Component {
+export default class CartCheckListComponent extends Component {
   @service('shopping-cart') cart;
 }
