+++ app/components/checkout-main.hbs
@@ -1,10 +1,12 @@
 <main class="w-11/12 flex flex-wrap my-0 mx-auto justify-start">
-    <div class="name-products-list custom-min-width-320 w-7/12 flex flex-col justify-between items-center mr-8">
+    <div class="name-products-list custom-min-width-320 w-7/12 flex flex-col justify-start items-center mr-8">
         {{#each this.cart.cartItems as |item|}}
             <CartProduct @product={{item}}/>
         {{/each}}
     </div>
     <div class="name-checkout-list w-4/12 custom-min-width-365">
-        <CartCheckList />
+        <CartCheckList 
+            @totalPrice={{this.cart.totalCartPrice}}
+         />
     </div>
 </main>
