+++ app/components/product.hbs
@@ -10,10 +10,10 @@
     </div>
     <div class="flex justify-end items-end w-2/4">
         <AddBtn
-          @quantity={{this.quantity}}
-          @deleteFromCart={{this.deleteFromCart}}
+          @product={{@product}}
+          {{!-- @deleteFromCart={{this.deleteFromCart}}
           @addToCart={{this.addToCart}}
-          @isconIsShown={{this.isconIsShown}}
+          @isconIsShown={{this.isconIsShown}} --}}
         />
     </div>
   </div>
