+++ app/services/shopping-cart.js
@@ -8,16 +8,21 @@ export default class ShoppingCartService extends Service {
   add(item) {
     const existingItem = this.cartItems.findBy('id', item.id);
     if (existingItem) {
-      const index = this.cartItems.indexOf(existingItem); // If an item with the same id exists, delete it and then replace it with the new item
-      this.cartItems.removeAt(index);
+      const newItem = {
+        ...existingItem,
+        quantity: existingItem.quantity + 1,
+      };
+      this.cartItems.removeObject(existingItem);
+      this.cartItems.pushObject(newItem);
+    } else {
+      this.cartItems.pushObject(item); // Push the new item into the cartItems array
     }
-    this.cartItems.pushObject(item); // Push the new item into the cartItems array
   }
 
   remove(id) {
     // gets id from product card
     const existingItem = this.cartItems.findBy('id', id);
-  
+
     if (existingItem) {
       if (existingItem.quantity > 1) {
         // If the quantity is greater than 1, create a new item with decremented quantity
@@ -42,7 +47,9 @@ export default class ShoppingCartService extends Service {
 
   get totalCartPrice() {
     let priceAmmount = 0;
-    this.cartItems.forEach((item) => (priceAmmount += item.discountedTotalPrice));
+    this.cartItems.forEach(
+      (item) => (priceAmmount += item.discountedTotalPrice)
+    );
     return priceAmmount;
   }
 }
